adding 2 person household cdap
This cdap is proving to be quiet tricky (mainly the 'coordinated' part).  The unit of decision is actually households rather than persons (coordinated across households) so I made that change.  Now some variables need to be contructed for households that weren't there - both age and sex need to be carried from the persons dataframe to the households dataframe for one person households.  And the ptype for households is a concatenation of the ptype for constituent persons.  After making these changes, things are beginning to work a bit more sanely.  The segmentation by household size is currently within the Python rather than in YAML.  I tend to like this because it gives me complete flexibility but we'll see what the group thinks.  As always, more to come...